+++ benchmarks/shared/src/main/scala/monix/benchmarks/ObservableConcatMapBenchmark.scala
@@ -60,7 +60,7 @@ class ObservableConcatMapBenchmark {
 
   @Benchmark
   def onePerCycle(): Long = {
-    val stream = Observable.range(0, size).flatMap { x =>
+    val stream = Observable.range(0, size.toLong).flatMap { x =>
       Observable.now(x + 1)
     }
     sum(stream)
@@ -68,7 +68,7 @@ class ObservableConcatMapBenchmark {
 
   @Benchmark
   def morePerCycle(): Long = {
-    val stream = Observable.range(0, size / 1000).flatMap { x =>
+    val stream = Observable.range(0, size.toLong / 1000).flatMap { x =>
       Observable.range(x, x + 1000)
     }
     sum(stream)
